make interface for api client
